Support IP layer mode on FreeBSD and merge Mac with BSD send code
