+++ src/send-bsd.h
@@ -9,6 +9,7 @@
 #ifndef ZMAP_SEND_BSD_H
 #define ZMAP_SEND_BSD_H
 
+#include <sys/param.h>
 #include <sys/types.h>
 #include <sys/time.h>
 #include <sys/ioctl.h>
@@ -23,10 +24,6 @@
 #ifdef ZMAP_SEND_LINUX_H
 #error "Don't include both send-bsd.h and send-linux.h"
 #endif
-#ifdef ZMAP_SEND_MAC_H
-#error "Don't include both send-bsd.h and send-mac.h"
-#endif
-
 
 int send_run_init(UNUSED sock_t sock)
 {
@@ -34,12 +31,40 @@ int send_run_init(UNUSED sock_t sock)
 	return EXIT_SUCCESS;
 }
 
-int send_packet(sock_t sock, void *buf, int len, UNUSED uint32_t idx)
+int send_packet(sock_t sock, void *buf, int len, UNUSED uint32_t retry_ct)
 {
-	return write(sock.sock, buf, len);
+	if (zconf.send_ip_pkts) {
+		struct ip *iph = (struct ip *)buf;
+
+#if defined(__APPLE__) || (defined(__FreeBSD__) && __FreeBSD_version < 1100030)
+		// Early BSD's raw IP sockets required IP headers to have len
+		// and off fields in host byte order, as they were being byte
+		// swapped on the way out.  Getting byte order wrong would
+		// result in EINVAL from sendto(2) below.
+		// Most modern BSD systems have fixed this and removed the byte
+		// swapping on raw IP sockets, while some, notably macOS, still
+		// require header fields in host byte order.
+		// See ip(4) for details on byte order requirements of raw IP
+		// sockets.
+		if (retry_ct == 0) {
+			iph->ip_len = ntohs(iph->ip_len);
+			iph->ip_off = ntohs(iph->ip_off);
+			iph->ip_sum = 0;
+		}
+#endif
+
+		struct sockaddr_in sai;
+		bzero(&sai, sizeof(sai));
+		sai.sin_family = AF_INET;
+		sai.sin_addr.s_addr = iph->ip_dst.s_addr;
+		return sendto(sock.sock, buf, len, 0, (struct sockaddr *)&sai, sizeof(sai));
+	} else {
+		return write(sock.sock, buf, len);
+	}
 }
 
-// BSD handles sockets differently than linux, and it seems non-trivial to port the linux code. Leaving this code that wraps the basic send_packet for now.
+// macOS doesn't have sendmmsg as of Sonoma. Since we want a uniform interface, we'll emulate the send_batch used in Linux.
+// FreeBSD does have sendmmsg, but it is a libc wrapper around the sendmsg syscall, without the perf benefits of sendmmsg.
 // The behavior in sendmmsg is to send as many packets as possible until one fails, and then return the number of sent packets.
 // Following the same pattern for consistency
 // Returns - number of packets sent
@@ -53,7 +78,7 @@ int send_batch(sock_t sock, batch_t* batch, int retries) {
 	int rc = 0;
 	for (int packet_num = 0; packet_num < batch->len; packet_num++) {
 		for (int retry_ct = 0; retry_ct < retries; retry_ct++) {
-			rc = send_packet(sock, ((void *)batch->packets) + (packet_num * MAX_PACKET_SIZE), batch->lens[packet_num], 0);
+			rc = send_packet(sock, ((uint8_t *)batch->packets) + (packet_num * MAX_PACKET_SIZE), batch->lens[packet_num], retry_ct);
 			if (rc >= 0) {
 				packets_sent++;
 				break;
@@ -63,16 +88,15 @@ int send_batch(sock_t sock, batch_t* batch, int retries) {
 			// packet couldn't be sent in retries number of attempts
 			struct in_addr addr;
 			addr.s_addr = batch->ips[packet_num];
-			char addr_str_buf
-			    [INET_ADDRSTRLEN];
+			char addr_str_buf[INET_ADDRSTRLEN];
 			const char *addr_str =
 			    inet_ntop(
 				AF_INET, &addr,
 				addr_str_buf,
 				INET_ADDRSTRLEN);
 			if (addr_str != NULL) {
-				log_debug( "send", "send_packet failed for %s. %s", addr_str,
-				    strerror( errno));
+				log_debug( "send-bsd", "send_packet failed for %s. %s", addr_str,
+					   strerror( errno));
 			}
 		}
 	}
